refactor(brand-scraper): extract website parser
