+++ apps/server/api/src/services/brand-scraper/brand-scraper.module.ts
@@ -1,8 +1,9 @@
 import { BrandScraperService } from '@api/services/brand-scraper/brand-scraper.service';
+import { BrandWebsiteParserService } from '@api/services/brand-scraper/brand-website-parser.service';
 import { Module } from '@nestjs/common';
 
 @Module({
   exports: [BrandScraperService],
-  providers: [BrandScraperService],
+  providers: [BrandScraperService, BrandWebsiteParserService],
 })
 export class BrandScraperModule {}
